feat: multicore scheduler with configured units
- Basic Round Robin scheduler on both cores with a predefined config
given by `scheduler.Config`.
- Each core defines a the number of quantum at which it is meant to
switch context to the next task in the queue.
- SysTick configured are specific to a core, hence if a core needs
to be interrupted, the systick needs to be configured on that core.

This shows that there is no need for having a "unit" (or quantum
time) of execution for a core to decide if it has to move to the
next task assuming that there is only a single SysTick Interrupt
for both cores.